firmware/2lib/2misc: Prepare for a extra reboot
To address a firmware bug b/180686277, a mitigation is added which
introduces an additional EC reset. But when both the kernel partitions
are corruption, this mitigiation introduces a reboot loop. To fix the
reboot loop, a firmware branch only vboot API is introduced to prepare
for the extra reboot. This work-around API stores the FW_TRIED and
FW_RESULT data into vboot's NV storage before triggering a reboot.

BUG=b:184388049
BRANCH=dedede
TEST=Build and boot to OS in Drawlat with a RO 13606.52.0 and RW
13606.199.0 + the fix. Ensure that the firmware_CorruptBothKernelAB.dev
test passes instead of boot looping.

Change-Id: Id53b9131034fcad6caeb3231c1c92d020b766338
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>