futility: updater: correct the rootkey in manifest.json
For customization projects, the bios image should be patched by
model first, then the gbb got from the image could be correct.

Reason:
In the function, patch_image_by_model, it will apply the rootkey,
vblock_a and vblock_b of the customization project to the gbb part
of the image.

This change moves the function "find_gbb" to the position after the
function "patch_image_by_model" is called.

BUG=b:188861251
BRANCH=none
TEST=build

Change-Id: I5d80405a2a3b0ecc45e86ca3322b328479b83702
Signed-off-by: Isaac Lee <isaaclee@google.com>