Avoid using malloc() directly in firmware
This library is supposed to stand alone, so has avoided making use of
standard library functions for the most part. Recent development has
resulted in malloc() being used, presumably in cases where there is
no chance of a memory leak.

However the bootloader may wish to implement this differently from the
C-library version, track the memory usage of vboot, or control whether
valgrind sees these calls. Follow the existing approach of using a
vb2ex...() function for this, to permit this flexibility.

Note that vbex_free(NULL) is defined as a nop. Adjust the code for this
while we are here.

This partially reverts 559a110f ("vboot: use malloc and free directly").

BUG=b:191646387
BRANCH=none
TEST=FEATURES=test emerge-volteer depthcharge

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I1c330f7e7ada7e6f4d0ec0da97dba3658aff460a