host: return a static HWID on REVEN devices
As b/195007267 describes, the normal call stack for `crossystem`
library to get the HWID string doesn't work on REVEN devices because
it's impossible to place the HWID into the firmware.

In order to make `crossystem` capable of returning a valid HWID string
on those devices, this CL introduces yet another layer of customization
for the host library called `class`.  This customization layer contains
nothing for usual projects.  But for `REVEN`'s build, it overrides
the getter function for `HWID` to return the pre-defined value.

BUG=b:195007267
TEST=manual test
BRANCH=none

Signed-off-by: Yong Hong <yhong@chromium.org>
Cq-Depend: chromium:3060798
Change-Id: Ie6238828a09f43390cc9d61ed33f76be7fe7454c