2nvstorage: Clear FIRMWARE_SETTINGS_RESET flag on a clean boot
The FIRMWARE_SETTINGS_RESET flag was intended to show when nvdata was
forcibly reset due to CRC failures. However, in practice it is useless
because it never gets cleared again and just stays set forever. This
patch fixes that by clearing it on every normal boot (i.e. a boot where
the exisiting nvdata was valid and didn't need to be reset).

BRANCH=None
BUG=None
TEST=Booted CoachZ, confirmed flag got cleared.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I8c88ea134bf92ccb5f1d9a710d86c5ec5198def8