vboot/ui: Just boot the one altfw if there's only one
Sometimes on my Chrome OS devices I like to use MrChromebox to install
alternate firmware. I'll use this to run other distros on a Chromebook I've
repurposed as a backup server. It used to be I could set the GBB flags
to boot the alternate firmware with a short delay, and this Chromebook
sitting in a closet would reboot reliably unattended.

With the new firmware UI, I get stuck in a "select alternate bootloader"
screen. This breaks my workflow, as now the device is hung in firmware
waiting for me to make what amounts to a non-choice.

Let's add a shortcut such that if there's only one alternate firmware,
we'll just attempt to boot it instead of presenting the UI screen. If
that attempt fails or there is more than one alternate firmware, then
present the UI screen with our attempted choice selected.

BUG=None
TEST=Run fw with altfw populated on dedede device.
BRANCH=main

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: Id2f5e7309af250f3c936de83a42a48b656ae342d