vboot/ui: Show error box on internal boot failure
CL:2917623 changes the behavior when failing to boot from internal disk
in developer UI. However, it causes the device to potentially enter a
reboot loop (b/197216832).

Instead of either triggering recovery or returning an error from the UI
loop, change the behavior to showing an error message in a dialog box
and staying in the UI. This is similar to the behavior when booting from
an invalid external disk, where an error screen will be shown to the
user.

BUG=b:197216832, b:197911901
TEST=make runtests
BRANCH=none

Change-Id: Ic5ace55fc8b93ba0836e5722b7c5011fd490c35e
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>