2lib/2api: Prevent unnecessary early returns in vb2api_fw_phase1
vb2_context is introducing a new filed to sotre the current boot mode
which shoulde be decided in the end of vb2api_fw_phase1. Therefore, any
early returns in vb2api_fw_phase1 will cause the boot mode undefined and
we will need extra efforts to handle this exception.

In conclusion, refactor and prevent most of the early returns except for
the VB2_ERROR_API_PHASE1_SECDATA_REBOOT case which will directly trigger
a reboot and do not need the boot mode.

BUG=b:181931817
BRANCH=none
TEST=CC=x86_64-pc-linux-gnu-clang;
     make clean && make runtests

Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: I24fd02e31b63a44bad93071965420c0303e26fbe