2lib/2misc: Set ctx flags VB2_CONTEXT_RECOVERY_MODE if vb2api_fail()
If we called vb2api_fail() and reach the last section, it means we
failed before choosing sections or both slots failed and we need to go
to recovery.

By refactoring and removing the early return in vb2api_fw_phase1 in
CL:2944250, we expect it catches any vb2api_fail() calls at the last
section `Return error if recovery is needed` and returns
VB2_ERROR_API_PHASE1_RECOVERY. However, we did not set that flag in
vb2api_fail() and will cause the function return VB2_SUCCESS instead.

Therefore, we decide to set the ctx flags along with setting the nv data
recovery reason and recovery code.

BUG=b:181931817
BRANCH=none
TEST=CC=x86_64-pc-linux-gnu-clang;
     make clean && make runtests

Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: Id1f6fc0eefaf3ecc32c93427bf089d34e5424b49