strip_boot_from_image: use sfill instead of fstrim
fstrim relies on the underlying device supporting it and on our
instances we're on lvm devices. Since we're fstrim'ing a mounted
loopback the ability to call fstrim is inherited by the parent device.
Something has changed in the kernel (see bug) that stopped us from
making the trim call on the loopback partition.

sfill with these options should accomplish the same thing (single
write with all zeros) as well as cleaning up inode space.

BRANCH=main
BUG=b:200038130
TEST=signer full tests with this commit.
Signed-off-by: engeg <engeg@google.com>

Change-Id: I8c71adfd59c11b5142aa367fb20222fc4b03a2ba