make_dev_ssd.sh: Do not write to active dm-verity root partition
Writing to the root partition while there is a dm-verity device
active on top of it is dangerous: if the dm-verity devices reads
the data from disk again, it will panic and invalidate the associated
kernel partition. This can innocuously happen if anything tries to
scan a device and can be hard to trace.

Instead of doing this, let developers remount the partition as
read-write on next boot.

BUG=b:207167831
TEST=none
BRANCH=none

Change-Id: I731ca5a523b76e035d962243edec4fc6667b102b
Signed-off-by: Sarthak Kukreti <sarthakkukreti@google.com>