futility: create powerd lock file when updating firmware
This functionality existed in flashrom for several years to stop the
power manager from suspending or shutting down the device while firmware
was being written to the flash.

It is being moved to futility now to facilitate migration to libflashrom
and reduce tech debt in flashrom.

While flashrom only locked during write and erase operations, futility
now locks for the entire firmware update process, including flash reads.
Writing firmware is usually slowest part of the process, so the lock is
not held for much longer than is strictly necessary and suspending while
futility is running is undesirable anyway.

BUG=b:207787495
BRANCH=none
TEST=Ran `futility update -i image.bin` and monitored lock files with:
  `cd /run/lock/power_override/ && \
   watch -n 0.1 'echo; ls -lah; echo; cat *.lock; echo; pgrep futility'`
TEST=Checked PID in lock file matched futility's PID.
TEST=Checked lock file was readable by powerd (i.e. '-rw-r--r--').
TEST=Checked lock file was deleted at futility exit.

Cq-Depend: chromium:3312440
Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: Iffcd835624bc73b43f0a8b96435dba5ff25e2bc9