2lib/2auxfw_sync: Handle VB2_REQUEST* error codes
In normal/secure mode, display is not always available. Hence aux
firmware updates request reboot to initialize the display so that slow
firmware update screen can be displayed. Since this error is not
handled, the concerned firmware slot is marked as failed. This leads to
devices in the field not picking FW updates and falling back to old FW
slot. Handle VB2_REQUEST* error codes, so that the device reboots to
initialize the display.

BUG=b:210127173
BRANCH=dedede, volteer, zork, trogdor
TEST=Build and boot to OS in Drawlat. Ensure that the AP + EC + ME + PD
FW updates are applied and the system was able to boot to OS
successfully - https://paste.googleplex.com/4966135716904960.

Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Change-Id: I2b7bdd830a1141ee59e25204bf4f77adba58b679