Revert "vboot_reference/futility: Inline get_host_wp()"
This reverts commit 5d40740b711a34a006355b877b3ba7303dbf743d.

Changing the write path in futility to use libflashrom had the side
effect of skipping the code in cli_classic that created a powerd lock
file. This could result in powerd suspending during a firmware update,
corrupting the RW firmware.

Revert back to subprocessing flashrom for M98.

BUG=b:214485250
BRANCH=release-R98-14388.B
TEST=emerge-grunt vboot_reference

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: Iffead23334e94d9b957e5179e5a66e0d9d5d0fb5