Revert "vboot_reference/futility: pipe log level though"
This reverts commit 2821ee7afac6b1a5d7998246f3a8026e914ef2e7.

Changing the write path in futility to use libflashrom had the side
effect of skipping the code in cli_classic that created a powerd lock
file. This could result in powerd suspending during a firmware update,
corrupting the RW firmware.

Revert back to subprocessing flashrom for M98.

BUG=none
BRANCH=release-R98-14388.B
TEST=emerge-grunt vboot_reference

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: Ic7dd82ac70d9f7047dd6f30450f8bb4a69a6f96c