Revert "vboot_reference/futility: Port R path to using libflashrom"
This reverts commit 0cd8ad2c330ee40beb54c2ce3def4ba6ea84fa3c.

Changing the write path in futility to use libflashrom had the side
effect of skipping the code in cli_classic that created a powerd lock
file. This could result in powerd suspending during a firmware update,
corrupting the RW firmware.

Revert back to subprocessing flashrom for M98.

BUG=b:214485250
BRANCH=release-R98-14388.B
TEST=emerge-grunt vboot_reference

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I3909e218258853a99753f38e27ca1b1b4fe54964