Revert "vboot_reference/futility: set layout when using region"
This reverts commit 91ffb514310bbc1fff166a049d9a40d11850f31.

Changing the write path in futility to use libflashrom had the side
effect of skipping the code in cli_classic that created a powerd lock
file. This could result in powerd suspending during a firmware update,
corrupting the RW firmware.

Revert back to subprocessing flashrom for M98.

BUG=none
BRANCH=release-R98-14388.B
TEST=emerge-grunt vboot_reference

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I6231564a6708033c5bd7919dbf963d44e63c73cb