Revert "futility: updater_setup_config() allow for non-host prog"
This reverts commit c2d0cc1d30103e9a9b62db03a88489ef47fca619.

Changing the write path in futility to use libflashrom had the side
effect of skipping the code in cli_classic that created a powerd lock
file. This could result in powerd suspending during a firmware update,
corrupting the RW firmware.

Revert back to subprocessing flashrom for M98.

BUG=none
BRANCH=release-R98-14388.B
TEST=emerge-grunt vboot_reference

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: Ie1551db7e837e53b5a26d4686550984b0b9152c8