Revert "futility: always create a tempfile in updater_apply_white_label()"
This reverts commit 0457e74c956b320e1ed24fee469f951521a3e18a.

Changing the write path in futility to use libflashrom had the side
effect of skipping the code in cli_classic that created a powerd lock
file. This could result in powerd suspending during a firmware update,
corrupting the RW firmware.

Revert back to subprocessing flashrom for M98.

BUG=none
BRANCH=release-R98-14388.B
TEST=emerge-grunt vboot_reference

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: If8712166b4bed533d780202401812e2823e5ac81