Revert "vboot_reference/futility: Port W path to using libflashrom"
This reverts commit 483dff64346fd30224df5b66bf76124aeea7bb12.

Changing the write path in futility to use libflashrom had the side
effect of skipping the code in cli_classic that created a powerd lock
file. This could result in powerd suspending during a firmware update,
corrupting the RW firmware.

Revert back to subprocessing flashrom for M98.

BUG=b:214485250
BRANCH=release-R98-14388.B
TEST=emerge-grunt vboot_reference

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I81ba6e45aa926d7ff930d280d2f4a5ef31c59fa2