vboot: Add --programmer and --servo flags to {get,set}_gbb_flags.sh.
Allow custom programmers to be used instead of host for getting and
setting GBB flags via a --programmer/-p flag similar to the same flag
for futility. Also support --servo with the same semantics as the
futility flag - detecting the programmer to use from the current servod
instance mirroring the logic in futility. Only CCD and servo micro are
supported, however.

When reading, include the FMAP section as well as the GBB section. For
writes over CCD, this can save up to 1.2s, by avoiding rescanning for
the FMAP.

BUG=None
TEST={get,set}_gbb_flags.sh --servo with and without servod running
BRANCH=None

Change-Id: Iecedf4c3d0cad6923aed4405ef4a72910f3f9f05
Signed-off-by: Sam McNally <sammc@chromium.org>