futility: updater: accumulate changes to prevent multiple partial write
The 'partial write' (flashrom -i SECTION) takes much longer than we
expected because it may need to read back (even verify) the whole image
for each run - and this will make RW recovery super long because it has
4 sections to update. As a result, we want to first create a temporary
firmware image, write all necessary changes, and then flash to the
system with only one flashrom command.

A new config entry 'direct_partial_write' is added so we can fallback
to the original behavior if needed (can be a quirk).

BUG=b:221137867
TEST=build and run test
BRANCH=None

Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Change-Id: Ifbc67327a02096e027c1e2025485ebb17645a71d