futility: flashrom_drv: merge parameters to a single flashrom_params
There may be more and more optional flags and parameters that we will
want to override (for example --noverify_all) from the futility programs
when calling the flashrom driver. Having a huge parameter list is not
easy for maintenance, so a new structure flashrom_params is introduced
for adding new parameters easily.

BUG=None
TEST=make; build and run test
BRANCH=None

Change-Id: Ic8a8cf01a5e57d19cf9b073af4491a36c4ed26d6
Signed-off-by: Hung-Te Lin <hungte@chromium.org>