host/lib/flashrom_drv: flashrom_write_image() should consume bounds
Fix memory corruption issue where a bounds is not passed when
indexing the regions array. As per the C spec:

(p3) Except when it is the operand of the sizeof operator, the _Alignof
operator, or the unary '&' operator, or is a string literal used to
initialize an array, an expression that has type "array of type" is
converted to an expression with type "pointer to type" that points to
the initial element of the array object and is not an lvalue. C11
Standard - 6.3.2.1 Other Operands - Lvalues, arrays, and function
designators(p3)

BUG=b:207808292
BRANCH=none
TEST=emerge-octopus and cros deploy'ed

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Change-Id: I54549cba00b5ef629720e89e9791aa56696f9c7d