futility: updater: refactor [write|load]_system_firmware
The 'write_system_firmware' and 'load_system_firmware' have too many
parameters that are already contained in the updater_config.
To simplify the interfaces, we should just pass the updater_config to
[write|load]_system_firmware, and merge the write_firmware_sections
(only handling emulation and deciding the diff image) to the
write_system_firmware.

Also moved the utility functions only used by *_system_firmware (for
example is_the_same_programmer and emulate_write_firmware) to the
updater_utils.c.

The emulate_write_firmware is also revised to handle a list of sections
directly (so we don't need to handle that in write_system_firmware).

BUG=None
TEST=make; build and run test.
BRANCH=None

Change-Id: I4d123d12e8057da82d6c301899472e3773f3266e
Signed-off-by: Hung-Te Lin <hungte@chromium.org>