scripts/image_signing: Add ensure_amd_psp_flags script
Currently there is no script to ensure that AMD PSP flags are set
correctly in a firmware image. This commit adds ensure_amd_psps_flags to
handle that functionality. The script can check that certain flags are
set as well as checking that certain flags are not set. The amdfwread
tool is currently under review on coreboot.

BRANCH=none
BUG=b:202397678
TEST=Ran script with ChromeOS image, verify that it responds correctly
to PSP flag values

Signed-off-by: Robert Zieba <robertzieba@google.com>
Change-Id: Ie0864544b9b97704ee901d893b4d833c1ab068b9