2lib: Remove NO_BOOT condition for manual recovery
When SW sync is disabled (ie: during early development), we are unable
to boot into recovery at all (we end up booting into the broken screen
all the time instead).  It seems that NO_BOOT is triggered when the RW
image in the EC does not match the EC RW in the AP firmware (which is
also very common during early development).  This results in ALWAYS
entering the broken screen when recovery is triggered and SW sync is
disabled.

Not only this, if a hacker is able to modify the EC RW and AP RW
images so that AP RW can't verify the EC RW image anymore, then the
user will never be able to boot into recovery and recovery their
device.  Instead they'll only be able to boot into the broken screen.

BUG=b:216317864
BRANCH=None
TEST=run dut-control power_state:rec and make sure that we can boot
     into recovery and boot OS image from a USB.

Signed-off-by: Shelley Chen <shchen@google.com>
Change-Id: I95a6b710891fa2803412292483585438782d7660