futility: updater: split manifest implementation from updater_archive
The updater_archive has two functions: the implementation of archives,
and the creation of manifest. Since we are having more support of
archives in different format and the logic of manifests are getting more
complicated, it is better to move the manifest to its own file.

No functional changes, only internal refactoring.

BUG=None
TEST=make; run test
BRANCH=None

Change-Id: I01ff9561583357eac4c6add8a09384f1c4c9d671
Signed-off-by: Hung-Te Lin <hungte@chromium.org>