futility: gscvd: No output on success
Right now the `gscvd` command always prints the root key hash on
success. This is a bit confusing when the command is run as part of a
larger script and violates the common theme of UNIX commands to print no
output on success. This patch hides that functionality behind the
--debug flag.

Also, for better output stream hygiene, move all human-readable
informational output to the stderr stream.

BRANCH=none
BUG=none
TEST=none

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Idc1ad9e87ae39481271a74f617c41e685e67d59d